+++ CHANGELOG.md
@@ -15,6 +15,7 @@ that can be found in the LICENSE file. -->
 
 ### Fixes
 
+- Fix previewing selected assets behavior.
 - Use `PermissionRequestOption` as much as possible.
 
 ## 9.0.0-dev.2