+++ CHANGELOG.md
@@ -16,9 +16,10 @@ that can be found in the LICENSE file. -->
 
 ### Fixes
 
-- Fix previewing selected assets behavior.
+- Fix previewing selected assets' behavior.
 - Use `PermissionRequestOption` as much as possible.
 - Raise more errors for non-synced paths.
+- Fix the experience with `shouldRevertGrid`.
 
 ## 9.0.0-dev.2